// src/app/core/enums/shared-enums/filter-table.enum.ts
export enum InputType {
  TEXT_INPUT = "input-field",
  DROPDOWN_FIELD = "dropdown-field",
  DATE_PICKER = "date-picker",
}

export enum EvalType {
  TEXT = "text",
  COUNT = "count",
  BOOLEAN = "boolean",
  CUSTOM = "custom",
  DATE = "date",
}

export enum FilterType {
  INPUT = "input",
  DROPDOWN = "dropdown",
  DATE = "date",
  SELECT = "select",
}